Update to support Java 11 (test) builds, plus minor improvements
This is mainly a build update, with the Ant script varying the module
construction based on the choice of Java 8 vs 11 and even injecting the
relevant command-line differences into to manifest to account for the
differences in launching.
Had to change a test from the 'diffFirstLineOnly' comparison
(now removed) to the 'grepMessage' comparison, since the error messages
differ slightly but could be handled by grep.